requirements.txt
PyQt6
numpy
matplotlib
pyModbusTCP
PyYAML